Create CONTRIBUTING
**# Welcome! **
Thank you for contributing to CDC's open source projects! If you have any questions or doubts, don't be afraid to send them our way. We appreciate all contributions, and we are looking forward to fostering an open, transparent, and collaborative environment. 
	
Before contributing, we would also love for you to read our LICENSE, READ ME, and CODE OF CONDUCT files, which should also be in this repository. If you have any inquiries or questions not answered by the contents of this repository, feel free to send us an email at [chiic@cdc.gov](mailto: chiic@cdc.gov). 

# **Public Domain**
This project is in the public domain within the United States, and copyright and related rights in the work worldwide are waived through the [CC0 1.0 Universal public domain dedication.

](https://creativecommons.org/publicdomain/zero/1.0/) All contributions to this project will be released under the CC0 dedication. By submitting a pull request, you are agreeing to comply with this waiver of copyright interest.

# **Requesting Changes:**
Our pull request/merging process is designed to give the CDC Surveillance Team and others in our space an opportunity to consider and discuss any suggested changes. This policy affects all CDC spaces, both online and off, and all users are expected to abide by it. 

## **Open an issue in the repository**
If you don't have specific language to submit but would like to suggest a change or have something addressed, you can open an issue in this repository.  Team members will respond to the issue as soon as possible. 

## **Submit a Pull Request**
If you would like to contribute, please submit a pull request. 

In order for us to merge a pull request, it must: 

- Be at least seven days old. Pull Requests may be held longer if necessary to give people the opportunity to assess it. 

- Receive a +1 from a majority of team members associated with the request. If there is significant dissent between the team, a meeting will be held to discuss a plan of action for the pull request.