Update open_practices.md
This is how I would structure this document, feel free to adapt some of it or none of it. The biggest persisting issue I noticed was a lot of information was in multiple places and that just adds text so I tried to apply the DRY rule to it and make all the copy as brief and straightforward as possible.